cleaned up imports, css injection moved to components.py, misc
